# public/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"AnalyzeImageBody": {
				"properties": {
					"imageUrl": {
						"type": "string"
					},
					"prompt": {
						"type": "string"
					}
				},
				"required": [
					"imageUrl"
				],
				"type": "object",
				"additionalProperties": true
			},
			"GenerateTextBody": {
				"properties": {
					"prompt": {
						"type": "string"
					},
					"systemPrompt": {
						"type": "string"
					}
				},
				"required": [
					"prompt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"OpenaiGenerateImageBody": {
				"properties": {
					"prompt": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"n": {
						"type": "number",
						"format": "double"
					},
					"size": {
						"type": "string"
					},
					"response_format": {
						"type": "string"
					},
					"style": {
						"type": "string"
					},
					"quality": {
						"type": "string"
					},
					"uploadToCos": {
						"type": "boolean"
					},
					"includeBase64InResponse": {
						"type": "boolean"
					},
					"includeThumbnailInResponse": {
						"type": "boolean"
					},
					"saveToDb": {
						"type": "boolean"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"category_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"prompt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"GenerateImageByChatBody": {
				"properties": {
					"session_id": {
						"type": "string"
					},
					"prompt": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"messages": {
						"items": {
							"properties": {
								"content": {},
								"role": {
									"type": "string"
								}
							},
							"required": [
								"content",
								"role"
							],
							"type": "object"
						},
						"type": "array"
					},
					"model": {
						"type": "string"
					},
					"group": {
						"type": "string"
					},
					"stream": {
						"type": "boolean"
					},
					"uploadToCos": {
						"type": "boolean"
					},
					"temperature": {
						"type": "number",
						"format": "double"
					},
					"top_p": {
						"type": "number",
						"format": "double"
					},
					"frequency_penalty": {
						"type": "number",
						"format": "double"
					},
					"presence_penalty": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
...
			"BananaGenerateImageBody": {
				"properties": {
					"type": {
						"type": "string",
						"enum": [
							"text-to-image",
							"image-to-image"
						]
					},
					"model": {
						"type": "string"
					},
					"prompt": {
						"type": "string"
					},
					"aspectRatio": {
						"type": "string"
					},
					"imageUrls": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"imageUrl": {
						"type": "string"
					}
				},
				"required": [
					"model",
					"prompt"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "普通用户认证 Token"
			},
			"adminJwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "管理员认证 Token"
			}
		}
	},
	"info": {
		"title": "my-project",
		"version": "0.0.0",
		"contact": {}
	},
	"paths": {
...
		"/app/text-to-image/tasks/{taskId}": {
			"get": {
				"operationId": "GetUploadTaskStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"文生图"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/app/analyzeImage": {
			"post": {
				"operationId": "AnalyzeImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"OpenAI"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AnalyzeImageBody"
							}
						}
					}
				}
			}
		},
		"/app/generateText": {
			"post": {
				"operationId": "GenerateText",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"OpenAI"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GenerateTextBody"
							}
						}
					}
				}
			}
		},
		"/app/generateImage": {
			"post": {
				"operationId": "GenerateImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"OpenAI"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OpenaiGenerateImageBody"
							}
						}
					}
				}
			}
		},
		"/app/textToimageNew": {
			"post": {
				"operationId": "GenerateImageByChatCompletions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"OpenAI"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GenerateImageByChatBody"
							}
						}
					}
				}
			}
		},
		"/app/textToImage/tasks/{taskId}": {
			"get": {
				"operationId": "GetUploadTaskStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"OpenAI"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/app/banana-CreateImage": {
			"post": {
				"operationId": "GenerateImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Banana 文生图"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BananaGenerateImageBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}